MBL ADD : add is-voidable systeme to check if transaction can be cancel

+++ Controllers/TransactionController.py
@@ -3,6 +3,11 @@ from sqlmodel import Session, select
 from models.model import Transactions, BankAccount, User
 from database import engine
 
+def is_avoidableCheck(id_transaction, timer):
+    with Session(engine) as session:
+        transaction = session.exec(select(Transactions).where(Transactions.id == id_transaction)).first()
+        if timer > 5:
+            transaction.is_voidable = False
 
 def cancel_transaction(id_compteA, id_compteB, id_transaction):
     #Je recupere le montant de la transaction
@@ -13,25 +18,31 @@ def cancel_transaction(id_compteA, id_compteB, id_transaction):
 
         if not transaction:
             return {"Cette transaction n'existe pas"}
-    #Je rajoute le montant au solde du compteA
-        bankaccountA.solde+=transaction.amout
-    #Je soustret le montant au solde du compteB
-        bankaccountB.solde-=transaction.amout
-    #Je supprime la transaction
-        session.delete(transaction)
 
-    #Je sauvegarde les modifications dans la base
-        session.add(bankaccountA)
-        session.add(bankaccountB)
-        session.commit()
-        session.refresh(bankaccountA)
-        session.refresh(bankaccountB)
+        if transaction.is_voidable==True:
+        #Je rajoute le montant au solde du compteA
+            bankaccountA.solde+=transaction.amout
+        #Je soustret le montant au solde du compteB
+            bankaccountB.solde-=transaction.amout
+        #Je supprime la transaction
+            session.delete(transaction)
 
-        return {
-            "message": f"Transaction {id_transaction} annulée avec succès.",
-            "compteA_solde": bankaccountA.solde,
-            "compteB_solde": bankaccountB.solde
-        }
+        #Je sauvegarde les modifications dans la base
+            session.add(bankaccountA)
+            session.add(bankaccountB)
+            session.commit()
+            session.refresh(bankaccountA)
+            session.refresh(bankaccountB)
+
+            return {
+                "message": f"Transaction {id_transaction} annulée avec succès.",
+                "compteA_solde": bankaccountA.solde,
+                "compteB_solde": bankaccountB.solde
+            }
+        else:
+            return {
+                "message" : "Cette transaction n'est plus annulable"
+            }
 
 def show_transaction(id_compteA, id_compteB, id_transaction):
     with Session(engine) as session:
@@ -84,7 +95,8 @@ def send_money(id_compteA: int, id_compteB: int, amout: float):
         transaction = Transactions(
             id_compteA=id_compteA,
             id_compteB=id_compteB,
-            amout=amout
+            amout=amout,
+            is_voidable=True
         )
 
         session.add(transaction)
